data-collector/requirements.txt
flask
requests
mysql-connector-python
pillow
